either or, not both
